frontend/requirements.txt
PyQt6==6.7.0
pycryptodome==3.20.0
requests==2.32.3
python-dotenv==1.0.0
keyring==24.3.0
